feat: better structure for articles
