+++ frontend/src/widgets/article/GitHubContributors.tsx
