+++ frontend/src/widgets/article/index.ts
@@ -1,0 +1,1 @@
+export { ArticleWidget } from './ArticleWidget';
