+++ frontend/src/widgets/index.ts
@@ -1,3 +1,4 @@
+export { ArticleWidget } from './article';
 export { ButtonWidget } from './ButtonWidget';
 export { ProgressWidget } from './ProgressWidget';
 export { TooltipWidget } from './TooltipWidget';
